requirements.txt
fastapi
uvicorn
pyatv
pystray
Pillow
pydantic
websockets
